+++ R/oa_fetch.R
@@ -7,7 +7,7 @@
 oa_entities <- function() {
   c(
     "works", "authors", "institutions", "concepts",
-    "funders", "sources", "publishers"
+    "funders", "sources", "publishers", "topics"
   )
 }
 
@@ -346,7 +346,7 @@ oa_request <- function(query_url,
     if (verbose) cat("\nDownloading groups...\n|")
     while (!is.null(next_page)) {
       if (verbose) cat("=")
-      Sys.sleep(1 / 100)
+      Sys.sleep(1 / 10)
       query_ls[[paging]] <- next_page
       res <- api_request(query_url, ua, query = query_ls)
       data <- c(data, res[[result_name]])
@@ -391,7 +391,7 @@ oa_request <- function(query_url,
   res <- NULL
   for (i in pages) {
     if (verbose) pb$tick()
-    Sys.sleep(1 / 100)
+    Sys.sleep(1 / 10)
     next_page <- get_next_page(paging, i, res)
     query_ls[[paging]] <- next_page
     res <- api_request(query_url, ua, query = query_ls)
@@ -461,7 +461,7 @@ get_next_page <- function(paging, i, res = NULL) {
 #' See more at <https://docs.openalex.org/how-to-use-the-api/get-single-entities#the-openalex-id>.
 #' @param entity Character. Scholarly entity of the search.
 #' The argument can be one of
-#' c("works", "authors", "institutions", "concepts", "funders", "sources", "publishers").
+#' c("works", "authors", "institutions", "concepts", "funders", "sources", "publishers", "topics").
 #' If not provided, `entity` is guessed from `identifier`.
 #' @param options List. Additional parameters to add in the query. For example:
 #'